Added --export-configuration option which queries the remote iDRAC
for the current system settings and creates an xml file with the
exported attributes to the current directory.
This assumes that we want the export settings with the following
payload:
{
  "ExportFormat": "XML",
  "ShareParameters": {"Target": "ALL"},
  "IncludeInExport": "IncludeReadOnly,IncludePasswordHashValues"
}
Also included --clear-jobs option which will query for all current
jobs on the iDRAC queue and will subsequently clear the queue from
the job ids collected.